[FE-263 FE-264 FE-262] Fix bugs (#181)
* disbale edit

* add with empty

* fix parseInt bug

* fix bonus points

+++ frontend/src/components/Edition/Sections/AwardsSection/AwardsSection.tsx
@@ -87,7 +87,7 @@ export const AwardsSection = () => {
               ? {
                   ...selectedAward.award,
                   awardType: selectedAward.award.awardType.toUpperCase(),
-                  awardValue: parseInt(selectedAward.award.awardValue),
+                  awardValue: parseFloat(selectedAward.award.awardValue),
                   imageId: selectedAward.award.imageFile?.fileId ?? "-1",
                   categoryId: selectedAward.award.category.categoryId,
                 }
